shore up IMGME regex
The old regex was matching on way too many strings. e.g. 'time goes some speed' was matching because 'tIMe Goes soME' contains the right letters in the right order, and so the bot is returning an unasked for 'speed' image.

This new regex should only match strings that begin with 'image', although I threw in some question marks so that 'img' still works.
I also made the 'me' optional while I was at it. I'm not sure if that's a good idea or not but what the hell let's try it.